Add an video quality option
Add a video quality option to the Invidious preferences screen
This option controls the default quality of the instance
It has four values:

'dash'
'hd720'
'medium'
'small'

When you set a value, it is applied to all instances. So, when you change the instance, you will use the same quality you selected.