Add a captions selector option
This option allows you to choose captions for the video, you can select up to three captions.

The selected captions will appear based on their availability in the video, If the first chosen caption isn't available, the second one will be applied, and so on.

The captions will be applied to all instances of the video.